Move security checks to service class
